chore: move overlays to root in sync script
For RHTAP the Dockerfile needs to be in one of a limited set of
locations. The easiest to deal with is at root.

Moving Dockerfile to `./overlays` instead of `./overlays/cosign` so that
we can just have a generic move of all files under overlays and any
directory structure is copied, instead of picking files from specific
directories and moving them.

Signed-off-by: Lance Ball <lball@redhat.com>