docker: keep cosign-linux-amd64 - cp instead of mv
Instead of moving the binary to be named `cosign`, this change modifies
the Dockerfile so that the binary is copied to the new, shortened name,
leaving the original, more descriptive name. This simplifies its usage
in other images, such as the CLI download image where the original name
is desired.

Signed-off-by: Lance Ball <lball@redhat.com>