feat: make CLI server builds hermetic (#165)
* feat: make CLI server builds hermetic

This attempts to make the CLI server builds hermetic by removing the use
of `oc image extract`. This was the approach used in our TP2 release
because when the `ec` binary was added to `cosign`, `gitsign` and
`rekor-cli` the source build exceeded the pipeline results size limit
and caused build failures. We have since also changed how the download
image works and it is now separated into two images with the CLI
binaries, which are used as data for a generic HTTP server.

All of this is to say that we don't need `oc image extract` now because
the whole thing has been split into two images and should not trigger
size limits on pipeline results.

Signed-off-by: Lance Ball <lball@redhat.com>

* fixup: slueth out a sha from quay.io before snapshot exists

Signed-off-by: Lance Ball <lball@redhat.com>

* enable-hermetic-builds for clis

---------

Signed-off-by: Lance Ball <lball@redhat.com>
Co-authored-by: JasonPowr <20076537@mail.wit.ie>