chore: remove k8s v1.24 from e2e tests
The upstream has completely refactored the e2e tests. We have not pulled
those changes in due to the fact that they haven't been included in a
released version yet. However, since the last pull from upstream, k8s
v1.24 is no longer supported in the version of kind being used in the
tests. This commit removes that version from the tests but otherwise
leaves them unchanged.

Signed-off-by: Lance Ball <lball@redhat.com>