Update conformance test binary for signing config (#4367)
The signing config will now be provided on the sign path to test Rekor
v2, along with the trusted root for verifying bundles on the sign path.

This also adds support for providing a trusted root with
sign-blob/attest-blob. Currently, you can either provide just the CT log
key or we'll fetch the trusted root from the initialized TUF repo. Since
we are providing the trusted root for staging during signing now, this
PR also lets the user provide the trusted root they'll use for verifying
during signing.

Signed-off-by: Hayden <8418760+haydentherapper@users.noreply.github.com>