Support self-managed keys when signing with sigstore-go (#4368)
* Support self-managed keys when signing with sigstore-go

This creates a wrapper around the Keypair interface when a
SignerVerifier is provided for signing with KMS or any other provided
keys. This also retains support for --issue-certificate to request a
certificate for a managed key.

Fixes #4327

Signed-off-by: Hayden <8418760+haydentherapper@users.noreply.github.com>

* Add issue-certificate flags to attest and attest-blob

This is for uniformity with sign/sign-blob.

Signed-off-by: Hayden <8418760+haydentherapper@users.noreply.github.com>

* Refactor SignerFromKeyOpts to split Fulcio signer into its own method

Now, we can generate a SignerVerifier from a provided key without
mandating that we also request a Fulcio certificate when
"issue-certificate" is provided.

Signed-off-by: Hayden <8418760+haydentherapper@users.noreply.github.com>

* Use default options to load key material

Signed-off-by: Hayden <8418760+haydentherapper@users.noreply.github.com>

---------

Signed-off-by: Hayden <8418760+haydentherapper@users.noreply.github.com>